phase-2-web/backend/pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
